+++ common/src/index.ts
@@ -8,6 +8,6 @@ export * from './entities/normal-asks';
 export * from './entities/renewable-asks';
 export * from './entities/renewable-bids';
 export * from './entities/normal-bids';
-export * from "./entities/primary-asks";
-export * from "./entities/primary-bids";
-export * from "./entities/monthly-usages";
+export * from './entities/primary-asks';
+export * from './entities/primary-bids';
+export * from './entities/student-accounts/monthly-usages';
